PiratenKapern/src/test/java/test/game/Theme/NormalTest.java
package test.game.Theme;

import static org.junit.Assert.*;

import org.junit.Test;

import java.util.HashMap;
import entity.Dice;
import game.OneTurnScoreChange;
import game.Theme.Normal;

public class NormalTest {

	@Test
	public void test1() {
		//no coin/diamond bonus (row 2-10)
		System.out.println("test 1: ");
		Normal n = new Normal();
		HashMap<Dice.Face, Integer> map = new HashMap<Dice.Face, Integer>();
		int[][] counts = {
			//[0-5]:dice number [6]expect score [7]skulls from card
			{3,1,1,1,1,1,0,0},
			{4,0,0,0,2,2,0,0},
			{2,3,2,1,0,0,100,0},
			{2,4,1,1,0,0,200,0},
			{2,5,1,0,0,0,500,0},
			{2,6,0,0,0,0,1000,0},
			{1,7,0,0,0,0,2000,0},
			{0,8,0,0,0,0,4500,0}};
		for (int i = 0; i < 8; i++) {
			int totalSkull = counts[i][0];
			for (int j = 0; j < 6; j++) {
				if(totalSkull >=3) map.put(Dice.Face.values()[j], 0);
				else map.put(Dice.Face.values()[j], counts[i][j]);
			}
			map.put(Dice.Face.SKULL, totalSkull);
			System.out.println("["+i+"]except: "+counts[i][6]+ " result: "+n.scoreCalculation(map, 0).getChange());
			assertTrue(n.scoreCalculation(map,counts[i][7]).getChange() == counts[i][6]);
		}
	}
	
	@Test
	public void test2() {
		//with coin/diamond bonus (row 11-20)
		System.out.println("\ntest 2: ");
		Normal n = new Normal();
		HashMap<Dice.Face, Integer> map = new HashMap<Dice.Face, Integer>();
		int[][] counts = {
			//[0-5]:dice number [6]expect score [7]skulls from card
			{1,3,1,1,1,1,300,0},
			{0,4,0,0,2,2,1100,0},
			{0,5,0,1,2,0,700,0},
			{0,6,0,1,1,0,1100,0},
			{0,7,0,0,0,1,2600,0},
			{0,8,0,0,0,1,4600,0},
			{2,3,3,0,0,0,200,0},
			{1,3,4,0,0,0,300,0},
			{0,4,4,0,0,0,900,0}};
		for (int i = 0; i < 9; i++) {
			int totalSkull = counts[i][0];
			for (int j = 0; j < 6; j++) {
				if(totalSkull >=3) map.put(Dice.Face.values()[j], 0);
				else map.put(Dice.Face.values()[j], counts[i][j]);
			}
			map.put(Dice.Face.SKULL, totalSkull);
			System.out.println("["+i+"]except: "+counts[i][6]+ " result: "+n.scoreCalculation(map, 0).getChange());
			assertTrue(n.scoreCalculation(map,counts[i][7]).getChange() == counts[i][6]);
		}
	}

	@Test
	public void test3() {
		//coin/diamond oak + bonus (row 21-32)
		System.out.println("\ntest 3: ");
		Normal n = new Normal();
		HashMap<Dice.Face, Integer> map = new HashMap<Dice.Face, Integer>();
		int[][] counts = {
			//[0-5]:dice number [6]expect score [7]skulls from card
			{1,2,2,0,0,3,400,0},
			{1,2,1,0,0,4,600,0},
			{1,1,1,0,0,5,1000,0},
			{1,1,0,0,0,6,1600,0},
			{0,1,0,0,0,7,2700,0},
			{0,0,0,0,0,8,5300,0},
			{0,0,0,0,0,9,5400,0},
			{0,0,0,2,3,3,800,0},
			{0,0,0,1,4,3,1000,0},
			{0,0,0,0,4,4,1700,0},
			{0,0,0,0,3,5,1900,0}};
		for (int i = 0; i < 11; i++) {
			int totalSkull = counts[i][0];
			for (int j = 0; j < 6; j++) {
				if(totalSkull >=3) map.put(Dice.Face.values()[j], 0);
				else map.put(Dice.Face.values()[j], counts[i][j]);
			}
			map.put(Dice.Face.SKULL, totalSkull);
			System.out.println("["+i+"]except: "+counts[i][6]+ " result: "+n.scoreCalculation(map, 0).getChange());
			assertTrue(n.scoreCalculation(map,counts[i][7]).getChange() == counts[i][6]);
			assertTrue(n.scoreCalculation(map, counts[i][7]).getRange() == OneTurnScoreChange.Range.SELF); 
		}
	}
	
	@Test
	public void test4() {
		//skulls from card(row34-38)
		System.out.println("\ntest 4: ");
		Normal n = new Normal();
		HashMap<Dice.Face, Integer> map = new HashMap<Dice.Face, Integer>();
		int[][] counts = {
			//[0-5]:dice number [6]expect score [7]skulls from card
			{1,8,0,0,0,0,4500,1},
			{3,7,0,0,0,0,0,2},
			{1,5,0,0,2,1,1300,1},
			{2,6,1,1,0,0,1000,2}};
		for (int i = 0; i < 4; i++) {
			int totalSkull = counts[i][0];
			for (int j = 0; j < 6; j++) {
				if(totalSkull >=3) map.put(Dice.Face.values()[j], 0);
				else map.put(Dice.Face.values()[j], counts[i][j]);
			}
			map.put(Dice.Face.SKULL, totalSkull);
			System.out.println("["+i+"]except: "+counts[i][6]+ " result: "+n.scoreCalculation(map, counts[i][7]).getChange());
			assertTrue(n.scoreCalculation(map,counts[i][7]).getChange() == counts[i][6]);
		}
	}
}
